requirements.txt
faker
psycopg2-binary
astronomer-cosmos
sqlalchemy
fastapi
uvicorn
pydantic
python-dotenv
dbt-core
dbt-postgres